all handle features are implemented, including deleting editing and displaying QR, moreover QR generation on adding item is implemented

+++ merchant/merch.py
@@ -2,12 +2,14 @@ from flask import Blueprint
 from flask import Flask, Blueprint, render_template, request, redirect, flash
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager, UserMixin, login_user, login_required, current_user
+import qrcode
 import random
 import string
+import os
 
 #authStr = Blueprint('authStr', __name__, url_prefix='/store')
 
-app = Flask(__name__)
+app = Flask(__name__, static_url_path='/static')
 
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///login.db'
 app.config['SQLALCHEMY_BINDS'] = {'product' : 'sqlite:///product.db'}
@@ -21,16 +23,30 @@ class User(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(25), nullable=False, unique=True)
     password = db.Column(db.Integer, nullable=False, unique=True)
-    products = db.relationship('Product', backref='owner', lazy='dynamic')
+    products = db.relationship('Product', backref='user')
 
 class Product(db.Model):
     __bind_key__ = 'product'
-    id = db.Column(db.Integer, primary_key=True)
+    key = db.Column(db.String(15), nullable=False, primary_key=True)
     name = db.Column(db.String(50), nullable=False, unique=False)
     upc = db.Column(db.Numeric(12), nullable=True, unique=False)
-    description = db.Column(db.String(100), nullable=True, unique=False)
-    key = db.Column(db.String(15), nullable=False, unique=True)
+    description = db.Column(db.String(150), nullable=True, unique=False)
     ownerId = db.Column(db.Integer, db.ForeignKey('user.id'))
+    comments = db.relationship('Comment', backref='product')
+    feedbacks = db.relationship('Feedback', backref='product')
+
+class Comment(db.Model):
+    __bind_key__ = 'product'
+    id = db.Column(db.Integer, primary_key=True)
+    comment = db.Column(db.String(150), nullable=True)
+    rating = db.Column(db.Integer, nullable=False)
+    productKey = db.Column(db.String(15), db.ForeignKey('product.key'))
+
+class Feedback(db.Model):
+    __bind_key__ = 'product'
+    id = db.Column(db.Integer, primary_key=True)
+    comment = db.Column(db.String(150), nullable=True)
+    productKey = db.Column(db.String(15), db.ForeignKey('product.key'))
 
 @login_manager.user_loader
 def load_user(user_id):
@@ -60,16 +76,10 @@ def profile():
 @login_required
 def addItem():
     if request.method == 'POST':
-        pName = request.form['productName']
-        pUPC = request.form['upc']
-        pOwnerId = current_user.get_id()
-        pDescription = request.form['description']
-        pTokenKey = generateKey()
-        while (not (Product.query.filter_by(key=pTokenKey).first() is None)):
-            pTokenKey = generateKey()
-        newProduct = Product(name=pName, upc = pUPC, description=pDescription, key=pTokenKey, ownerId=pOwnerId)
-        db.session.add(newProduct)
-        db.session.commit()
+        name = request.form['productName']
+        upc = request.form['upc']
+        description = request.form['description']
+        addProduct(name, upc, description)
         return redirect('/addProduct')
     else:
         return render_template('add.html')
@@ -80,19 +90,54 @@ def manage():
     ownerItems = Product.query.filter_by(ownerId=current_user.get_id()).all()
     return render_template('manage.html', products=ownerItems)
 
-@app.route('/manage/delete/<int:id>')
+@app.route('/manage/delete/<id>')
 @login_required
 def delete(id):
-    item = Product.query.get_or_404(id)
-    db.session.delete(item)
-    db.session.commit()
+    deleteProduct(id)
     return redirect('/manage')
 
+@app.route('/manage/edit/<id>', methods=['GET', 'POST'])
+@login_required
+def edit(id):
+    item = Product.query.filter_by(key=id).first()
+    if request.method == 'GET':
+        return render_template('add.html', defName=item.name, defUpc=str(item.upc), defDesc=item.description)
+    else:
+        db.session.delete(item)
+        name = request.form['productName']
+        upc = request.form['upc']
+        description = request.form['description']
+        addProduct(name, upc, description)
+        return redirect('/manage')
 
+@app.route('/manage/displayqr/<id>')
+@login_required
+def displayQr(id):
+    return render_template('qr.html', qrSrc=qrPath(id))
 
 def generateKey():
     symbols = string.ascii_letters + string.digits
     key = ''.join(random.choice(symbols) for i in range(15))
     return key
 
-app.run()
+def deleteProduct(id):
+    item = Product.query.filter_by(key=id).first()
+    os.remove("static/images/"+id+'.png')
+    db.session.delete(item)
+    db.session.commit()
+
+def addProduct(nName, nUpc, nDescription):
+    productKey = generateKey()
+    while (not (Product.query.filter_by(key=productKey).first() is None)):
+        productKey = generateKey()
+    img = qrcode.make('http://192.168.0.13:5000/item/' + productKey)
+    img.save('static/images/'+productKey+'.png')
+    nOwnerId = current_user.get_id()
+    newProduct = Product(key=productKey, name=nName, upc = nUpc, description=nDescription, ownerId=nOwnerId)
+    db.session.add(newProduct)
+    db.session.commit()
+
+def qrPath(key):
+    return '/static/images/' + key + '.png'
+
+app.run(host='0.0.0.0')
